Move tests using _run_trial to test_optimize.py
