+++ tests/study_tests/test_optimize.py
@@ -1,14 +1,77 @@
+from typing import Callable
+from typing import Optional
 from unittest import mock
 
 import pytest
 
 from optuna import create_study
 from optuna import Trial
+from optuna import TrialPruned
 from optuna.study import _optimize
 from optuna.study._tell import _tell_with_warning
 from optuna.testing.storage import STORAGE_MODES
 from optuna.testing.storage import StorageSupplier
 from optuna.trial import TrialState
+
+
+@pytest.mark.parametrize("storage_mode", STORAGE_MODES)
+def test_run_trial(storage_mode: str) -> None:
+
+    with StorageSupplier(storage_mode) as storage:
+        study = create_study(storage=storage)
+
+        frozen_trial = _optimize._run_trial(study, lambda _: float("inf"), catch=())
+        assert frozen_trial.state == TrialState.COMPLETE
+        assert frozen_trial.value == float("inf")
+
+
+@pytest.mark.parametrize("storage_mode", STORAGE_MODES)
+def test_run_trial_automatically_fail(storage_mode: str) -> None:
+
+    with StorageSupplier(storage_mode) as storage:
+        study = create_study(storage=storage)
+
+        frozen_trial = _optimize._run_trial(study, lambda _: float("nan"), catch=())
+        assert frozen_trial.state == TrialState.FAIL
+        assert frozen_trial.value is None
+
+        frozen_trial = _optimize._run_trial(study, lambda _: None, catch=())  # type: ignore
+        assert frozen_trial.state == TrialState.FAIL
+        assert frozen_trial.value is None
+
+        frozen_trial = _optimize._run_trial(study, lambda _: object(), catch=())  # type: ignore
+        assert frozen_trial.state == TrialState.FAIL
+        assert frozen_trial.value is None
+
+        frozen_trial = _optimize._run_trial(study, lambda _: [0, 1], catch=())  # type: ignore
+        assert frozen_trial.state == TrialState.FAIL
+        assert frozen_trial.value is None
+
+
+@pytest.mark.parametrize("storage_mode", STORAGE_MODES)
+def test_run_trial_pruned(storage_mode: str) -> None:
+    def gen_func(intermediate: Optional[float] = None) -> Callable[[Trial], float]:
+        def func(trial: Trial) -> float:
+            if intermediate is not None:
+                trial.report(step=1, value=intermediate)
+            raise TrialPruned
+
+        return func
+
+    with StorageSupplier(storage_mode) as storage:
+        study = create_study(storage=storage)
+
+        frozen_trial = _optimize._run_trial(study, gen_func(), catch=())
+        assert frozen_trial.state == TrialState.PRUNED
+        assert frozen_trial.value is None
+
+        frozen_trial = _optimize._run_trial(study, gen_func(intermediate=1), catch=())
+        assert frozen_trial.state == TrialState.PRUNED
+        assert frozen_trial.value == 1
+
+        frozen_trial = _optimize._run_trial(study, gen_func(intermediate=float("nan")), catch=())
+        assert frozen_trial.state == TrialState.PRUNED
+        assert frozen_trial.value is None
 
 
 @pytest.mark.parametrize("storage_mode", STORAGE_MODES)
